# public/doc/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Essential life API",
        "description": "REST API consumed by Essential Life IOS and Android app",
        "contact": {
            "email": "ibnaalberunyibrahim@gmail.com"
        },
        "version": "2.1.0"
    },
    "paths": {
...
        "/{lang}/v2.1/users": {
            "get": {
                "tags": [
                    "V2.1-api-token"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\v2_1\\User\\UserController::show",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "sp",
                                "jp"
                            ]
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}